backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
sentence-transformers==2.2.2
ffmpeg-python==0.2.0
faiss-cpu==1.9.0.post1
openai-whisper @ git+https://github.com/openai/whisper.git
torch==2.6.0
openai>=1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
